[SofaKernel] Deep refactoring of DataTypeInfo.
