+++ SofaKernel/modules/SofaBaseMechanics/MechanicalObject.h
@@ -32,6 +32,8 @@
 #include <sofa/defaulttype/VecTypes.h>
 #include <sofa/defaulttype/RigidTypes.h>
 
+#include <sofa/helper/types/RGBAColor.h>
+
 #include <vector>
 #include <fstream>
