+++ SofaKernel/modules/SofaDefaultType/src/sofa/defaulttype/Color.h
@@ -19,26 +19,13 @@
 *                                                                             *
 * Contact information: contact@sofa-framework.org                             *
 ******************************************************************************/
-#ifndef SOFA_DEFAULTTYPE_COLOR_H
-#define SOFA_DEFAULTTYPE_COLOR_H
+#pragma once
 
-#include <functional>
-#include <limits>
-
-#include <sofa/helper/rmath.h>
 #include <sofa/helper/types/RGBAColor.h>
 
-#include <sofa/defaulttype/DataTypeInfo.h>
-
-
-namespace sofa
+namespace sofa::defaulttype
 {
-
-namespace defaulttype
-{
-    using sofa::helper::types::RGBAColor ;
+    [[deprecated("sofa::helper::types::RGBAColor is now part in sofa::helper::types::RGBAColor. Please update your code.")]]
+    typedef sofa::helper::types::RGBAColor RGBAColor;
 }
-}
-
-#endif
